Checkout latest tag if project_version is "latest"
In production purpose it is very useful to automatically checkout
the latest available tag. You can specify the tag each time, but it
is error prone and a very simple command will fetch the latest tag.

This cannot be performed outside the task, because the git repository
must exist before the tag can be retrieved. If the strategy is not
set to "git" or the version not to "master", all tasks are skipped.